requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
ollama>=0.2.0
python-dotenv>=1.0.0
psycopg2-binary
chromadb
sqlparse
faker>=24.0.0
streamlit
pandas
requests>=2.31.0
pydantic>=2.6.0
